build: complete Zensical and uv migration
