+++ .github/workflows/python-publish.yml
@@ -34,13 +34,10 @@ jobs:
         with:
           python-version-file: "pyproject.toml"
 
-      - name: Install the project
-        run: uv sync --all-extras --dev
       - name: Build package
-        run: uv build
-      # TODO: Change with `uv publish`
+        run: uv build --no-sources
       - name: Publish package
-        uses: pypa/gh-action-pypi-publish@ba38be9e461d3875417946c167d0b5f3d385a247 # v1.14.1
+        uses: pypa/gh-action-pypi-publish@dc37677b2e1c63e2034f94d8a5b11f265b73ba33 # v1.14.2
         with:
           user: __token__
           password: ${{ secrets.PYPI_API_TOKEN }}
