+++ .gitignore
@@ -4,9 +4,6 @@ playground/*
 !playground/*.py
 
 
-# lock files
-uv.lock
-
 # test
 junit.xml
 
@@ -158,7 +155,7 @@ venv.bak/
 # Rope project settings
 .ropeproject
 
-# mkdocs documentation
+# Zensical documentation output
 /site
 
 # mypy
