--- docs/blog/.authors.yml
@@ -1,6 +1,0 @@
-authors:
-  mathew:
-    name: Xiangzhuang Shen
-    description: Creator
-    avatar: https://avatars.githubusercontent.com/u/17157965
-    url: https://github.com/shenxiangzhuang
