--- docs/blog/index.md
@@ -1,1 +1,0 @@
-# Blog
