--- docs/blog/posts/about.md
@@ -1,20 +1,0 @@
----
-title: About LLM Glimpses
-draft: true
-date: 2024-03-22
-authors: [mathew]
-slug: about-llm-glimpses
-description: >
-    LLM Glimpses is a collection of essays that try to explain
-    the Large Language Model world in a simple way.
-categories:
-  - General
----
-
-# Why LLM Glimpses?
-现在，大型语言模型 (LLM, Large Language Model) 已经成为人工智能领域的热点。
-然而，对于大多数人来说，LLM 仍然是一个神秘的领域，或者说相对比较黑盒的领域。
-LLM Glimpses 的目标是通过简单的方式解释 LLM 世界，使更多的人能够理解 LLM 的基本概念。
-
-而 Glimpses 这个词，这里可以理解为“掠影”，“管窥”的含义。
-我们不会深入到 LLM 内部各种的细节，而是通过一些简单的例子，让读者能够对 LLM 有一个**直观且正确**的认识。
