+++ docs/changelog.md
@@ -1,1 +1,4 @@
---8<-- "CHANGELOG.md"
+# Changelog
+
+See the project [changelog](https://github.com/ai-glimpse/toyllm/blob/master/CHANGELOG.md)
+for release history.
