+++ docs/index.md
@@ -1,1 +1,7 @@
---8<-- "https://raw.githubusercontent.com/toy-ai/toyllm/master/README.md"
+# ToyLLM
+
+ToyLLM is a hands-on educational project for understanding and implementing
+Large Language Models from scratch.
+
+See the [project README](https://github.com/ai-glimpse/toyllm#readme) for
+installation instructions, usage examples, and the project structure.
