--- mkdocs.yml
@@ -1,112 +1,0 @@
-# Zensical reads this compatibility configuration while MkDocs plugin modules mature.
-site_name: ToyLLM
-repo_url: https://github.com/toy-ai/toyllm
-repo_name: ai-glimpse/toyllm
-site_description: Toy LLM
-site_author: Xiangzhuang Shen
-copyright: Copyright &copy; 2024 Xiangzhuang Shen
-
-# Page tree
-nav:
-  - Home: index.md
-  - Changelog: changelog.md
-  - Blog:
-      - blog/index.md
-
-theme:
-  name: material
-  icon:
-    repo: fontawesome/brands/github
-  language: en
-  palette:
-    - media: "(prefers-color-scheme: light)"
-      scheme: default
-      primary: teal
-      accent: deep purple
-      toggle:
-        icon: material/weather-sunny
-        name: Switch to dark mode
-    - media: "(prefers-color-scheme: dark)"
-      scheme: slate
-      primary: cyan
-      accent: deep purple
-      toggle:
-        icon: material/weather-night
-        name: Switch to light mode
-  features:
-    - content.code.copy
-    - content.action.edit
-    - content.action.view
-    # - announce.dismiss
-    - content.code.annotate
-    # - content.tabs.link
-    - content.tooltips
-    # - header.autohide
-    # - navigation.expand
-    - navigation.indexes
-    # - navigation.instant
-    # - navigation.prune
-    - navigation.sections
-    - navigation.tabs
-    # - navigation.tabs.sticky
-    - navigation.top
-    - navigation.tracking
-    - search.highlight
-    - search.share
-    - search.suggest
-    - toc.follow
-    # - toc.integrate
-
-docs_dir: docs
-
-extra_css:
-  - css/extra.css
-
-extra_javascript:
-  - js/config.js
-  - https://polyfill.io/v3/polyfill.min.js?features=es6
-  - https://cdn.jsdelivr.net/npm/mathjax@3/es5/tex-mml-chtml.js
-
-extra:
-  social:
-    - icon: fontawesome/brands/github
-      link: https://github.com/shenxiangzhuang/
-    - icon: fontawesome/brands/linkedin
-      link: https://www.linkedin.com/in/xiangzhuang-shen-a81825157/
-
-markdown_extensions:
-  - def_list
-  - pymdownx.tasklist:
-      custom_checkbox: true
-  - toc:
-      permalink: true
-  - pymdownx.arithmatex:
-      generic: true
-  - pymdownx.highlight:
-      linenums_style: pymdownx.inline
-  - pymdownx.superfences
-  - pymdownx.inlinehilite
-  - pymdownx.details
-  - pymdownx.snippets:
-      url_download: true
-  - pymdownx.tabbed:
-      alternate_style: true
-  - abbr
-  - admonition
-  - attr_list
-  - footnotes
-
-plugins:
-  - git-revision-date-localized:
-      enable_creation_date: true
-  - git-committers:
-      repository: ai-glimpse/toyllm
-      branch: master
-      token: !ENV [ "ORG_MKDOCS_GIT_COMMITTERS_APIKEY" ]
-  - blog:
-      draft: false
-  - search
-  - bibtex:
-      bib_file: "docs/references.bib"
-      csl_file: "docs/ieee.csl"
-      enable_inline_citations: false
