+++ prek.toml
@@ -1,33 +1,17 @@
 [[repos]]
-repo = "https://github.com/pre-commit/pre-commit-hooks"
-rev = "v6.0.0"
-
-[[repos.hooks]]
-id = "check-toml"
-
-[[repos.hooks]]
-id = "check-yaml"
-exclude = "mkdocs.yml"
-
-[[repos.hooks]]
-id = "end-of-file-fixer"
-
-[[repos.hooks]]
-id = "trailing-whitespace"
-exclude = ".+\\.csv"
-
-[[repos.hooks]]
-id = "mixed-line-ending"
-args = ["--fix=lf"]
+repo = "builtin"
+hooks = [
+    { id = "check-toml" },
+    { id = "check-yaml" },
+    { id = "end-of-file-fixer" },
+    { id = "trailing-whitespace", exclude = ".+\\.csv" },
+    { id = "mixed-line-ending", args = ["--fix=lf"] },
+]
 
 [[repos]]
-repo = "https://github.com/astral-sh/ruff-pre-commit"
-rev = "v0.16.0"
-
-[[repos.hooks]]
-id = "ruff"
-args = ["--config=pyproject.toml"]
-
-[[repos.hooks]]
-id = "ruff-format"
-args = ["--config=pyproject.toml"]
+repo = "local"
+hooks = [
+    { id = "ruff-check", name = "ruff check", entry = "uv run ruff check --config=pyproject.toml", language = "system", types_or = ["python", "pyi"] },
+    { id = "ruff-format", name = "ruff format", entry = "uv run ruff format --config=pyproject.toml", language = "system", types_or = ["python", "pyi"] },
+    { id = "ty", name = "ty check", entry = "uv run ty check", language = "system", pass_filenames = false, always_run = true, require_serial = true },
+]
