+++ src/toyllm/core/_generate.py
