+++ src/toyllm/core/_logits.py
