+++ src/toyllm/device.py
