+++ src/toyllm/sps/generate.py
