+++ src/toyllm/util/timer.py
