+++ zensical.toml
@@ -1,0 +1,3 @@
+[project]
+site_name = "ToyLLM"
+site_url = "https://ai-glimpse.github.io/toyllm/"
